+++ CHANGELOG.md
@@ -14,6 +14,7 @@ i.e. `<Major version>.<Minor version>.<Patch version>`
 - support centralized liberator
 - python lib broken upgrading (code rework)
 - introduce shield: separately control nftables with api/pubsub
+- remove access layer function
 
 ## [v0.8.0] - 2024-09-14
 - Support Debian 12 (bookworm)